Calender Code Structure verbeteren
