+++ web/src/features/hostdashboard/hostcalendar/utils/sortDates.js
@@ -1,0 +1,10 @@
+/**
+ * this function sorts the dates in the selectedDates array
+ *
+ * @param {[number,number][]} selectedDates
+ */
+function sortDates(selectedDates) {
+    selectedDates.sort((a, b) => a[0] - b[0]);
+}
+
+export default sortDates
